refactor: Move remaining tests to colocated structure
- Move tests/content/document/ to src/kurt/content/tests/document/
- Move tests/models/landing/ to src/kurt/models/landing/tests/
- Move tests/retrieval/ to src/kurt/models/staging/retrieval/tests/
- Merge retrieval conftest fixtures and fix import shadowing
- Remove empty tests/ directory
