+++ src/kurt/content/tests/document/test_document_with_analytics.py
