+++ src/kurt/models/staging/retrieval/tests/test_cag.py
@@ -17,8 +17,7 @@ from kurt.utils.retrieval import (
     search_entities_by_embedding,
 )
 
-# Import fixtures from main conftest
-from tests.conftest import reset_dbos_state, tmp_project  # noqa: F401
+# Fixtures tmp_project and reset_dbos_state are auto-discovered from conftest
 
 # ============================================================================
 # Unit Tests
