--- tests/retrieval/__init__.py
@@ -1,1 +1,0 @@
-"""Retrieval module tests."""
